+++ src/Modules/LearningTasks/Tutor.LearningTasks.Core/Domain/CaseStudy.cs
