+++ src/Modules/LearningTasks/Tutor.LearningTasks.Core/Domain/Standard.cs
@@ -1,17 +1,15 @@
 ﻿using Tutor.BuildingBlocks.Core.Domain;
 
-namespace Tutor.LearningTasks.Core.Domain
-{
-    public class Standard : Entity
-    {
-        public string Description { get; private set; }
-        public int MaxPoints { get; private set; }
+namespace Tutor.LearningTasks.Core.Domain;
 
-        public Standard(string description, int maxPoints)
-        {
-            Description = description;
-            MaxPoints = maxPoints;
-        }
+public class Standard : Entity
+{
+    public string Description { get; private set; }
+    public int MaxPoints { get; private set; }
+
+    public Standard(string description, int maxPoints)
+    {
+        Description = description;
+        MaxPoints = maxPoints;
     }
 }
-
