+++ src/Modules/LearningTasks/Tutor.LearningTasks.Infrastructure/Database/LearningTasksContext.cs
@@ -1,27 +1,25 @@
 ﻿using Microsoft.EntityFrameworkCore;
-using Tutor.LearningTasks.Core.Domain.Activites;
+using Tutor.LearningTasks.Core.Domain.Activities;
 
-namespace Tutor.LearningTasks.Infrastructure.Database
+namespace Tutor.LearningTasks.Infrastructure.Database;
+
+public class LearningTasksContext : DbContext
 {
-    public class LearningTasksContext : DbContext
+    public DbSet<Activity> Activities { get; set; }
+    public DbSet<Example> Examples { get; set; }
+
+    public LearningTasksContext(DbContextOptions<LearningTasksContext> options) : base(options) { }
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
     {
-        public DbSet<Activity> Activities { get; set; }
-        public DbSet<Example> Examples { get; set; }
+        modelBuilder.HasDefaultSchema("learningTasks");
+        modelBuilder.Entity<Activity>().Property(a => a.Guidance).HasColumnType("jsonb");
 
-        public LearningTasksContext(DbContextOptions<LearningTasksContext> options) : base(options) { }
+        modelBuilder.Entity<Activity>()
+            .HasMany(a => a.Examples)
+            .WithOne()
+            .HasForeignKey(e => e.ActivityId);
 
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            modelBuilder.HasDefaultSchema("learningTasks");
-            modelBuilder.Entity<Activity>().Property(a => a.Guidance).HasColumnType("jsonb");
-
-            modelBuilder.Entity<Activity>()
-                .HasMany(a => a.Examples)
-                .WithOne()
-                .HasForeignKey(e => e.ActivityId);
-
-            modelBuilder.Entity<Activity>().Property(a => a.Subactivities).HasColumnType("jsonb");
-        }
+        modelBuilder.Entity<Activity>().Property(a => a.Subactivities).HasColumnType("jsonb");
     }
 }
-
