+++ src/Modules/LearningTasks/Tutor.LearningTasks.Infrastructure/Database/Repositories/ActivityDatabaseRepository.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Tutor.BuildingBlocks.Infrastructure.Database;
-using Tutor.LearningTasks.Core.Domain.Activites;
+using Tutor.LearningTasks.Core.Domain.Activities;
 using Tutor.LearningTasks.Core.Domain.RepositoryInterfaces;
 
 namespace Tutor.LearningTasks.Infrastructure.Database.Repositories;
