+++ src/Modules/LearningTasks/Tutor.LearningTasks.Infrastructure/Tutor.LearningTasks.Infrastructure.csproj
@@ -13,6 +13,7 @@
 
   <ItemGroup>
     <Folder Include="Database\Repositories\" />
+    <Folder Include="Migrations\" />
   </ItemGroup>
 
 </Project>
